feat(#690): move vfs_hook_impls from core/ to services/
VFS hook implementations (DynamicViewerReadHook, AutoParseWriteHook,
TigerCacheRenameHook) are service-tier policy (S15/P17), not kernel
mechanism. The kernel provides the dispatch pipeline (VFSHookPipeline);
concrete hooks implement business policies like CSV filtering,
auto-parsing, and cache invalidation.

- Move core/vfs_hook_impls.py → services/vfs_hook_impls.py
- Move _register_vfs_hooks() from NexusFS.__init__ to factory/_wired.py
- Call register_vfs_hooks(nx) from factory orchestrator after construction
- Eliminates one core/ → services/ reverse dependency
