+++ src/nexus/services/vfs_hook_impls.py
